requirements.txt
pytest~=8.4.0
allure-pytest~=2.14.0
selenium~=4.33.0
allure-python-commons~=2.14.3